+++ example/lib/main.dart
@@ -17,7 +17,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-  String yourIdentificationId = 'YOUR_IDENTIFICATION_ID';
+  String yourIdentificationId = "cca07ee8-f547-45ad-b818-7baf6010446e";
   bool ondatoInitialized = false;
 
   @override